# Rounds chunk bytes to kb in tsdb and includes benchmarking script