--- a/tools/tsdb/tsdb-map/diff.sh
+++ b/tools/tsdb/tsdb-map/diff.sh
@@ -14,5 +14,5 @@ echo
 echo sizing:
 echo
 
-ls -lh $old
-ls -lh $new
+ls -lh "${old}"
+ls -lh "${new}"