ref https://github.com/grafana/loki/issues/5428. This gains us about ~5% space reduction over storing raw byte count.

Results:
```
benchmarks:

name                     old time/op    new time/op    delta
Query/match_ns-4           2.40µs ± 0%    2.42µs ± 0%   ~     (p=1.000 n=1+1)
Query/match_ns_regexp-4    2.42µs ± 0%    2.39µs ± 0%   ~     (p=1.000 n=1+1)

name                     old alloc/op   new alloc/op   delta
Query/match_ns-4            80.0B ± 0%     80.0B ± 0%   ~     (all equal)
Query/match_ns_regexp-4     80.0B ± 0%     80.0B ± 0%   ~     (all equal)

name                     old allocs/op  new allocs/op  delta
Query/match_ns-4             4.00 ± 0%      4.00 ± 0%   ~     (all equal)
Query/match_ns_regexp-4      4.00 ± 0%      4.00 ± 0%   ~     (all equal)

sizing:

-rw-r--r-- 1 owen users 40M Feb 25 10:47 /tmp/loki-scratch/loki-ops-tsdb
-rw-r--r-- 1 owen users 38M Feb 25 11:23 /tmp/loki-scratch/loki-ops-tsdb-experiment
```